rename modules folder to shader-modules
