+++ modules/effects/src/experimental/shader-modules/convolution.md
