+++ modules/effects/src/experimental/shader-modules/pack.md
