+++ modules/effects/src/index.js
@@ -1,32 +1,32 @@
 // Shader modules
 
 // glfx ADJUST shader modules
-export {default as brightnessContrast} from './modules/adjust-filters/brightnesscontrast';
-export {default as denoise} from './modules/adjust-filters/denoise';
-export {default as hueSaturation} from './modules/adjust-filters/huesaturation';
-export {default as noise} from './modules/adjust-filters/noise';
-export {default as sepia} from './modules/adjust-filters/sepia';
-export {default as vibrance} from './modules/adjust-filters/vibrance';
-export {default as vignette} from './modules/adjust-filters/vignette';
+export {default as brightnessContrast} from './shader-modules/adjust-filters/brightnesscontrast';
+export {default as denoise} from './shader-modules/adjust-filters/denoise';
+export {default as hueSaturation} from './shader-modules/adjust-filters/huesaturation';
+export {default as noise} from './shader-modules/adjust-filters/noise';
+export {default as sepia} from './shader-modules/adjust-filters/sepia';
+export {default as vibrance} from './shader-modules/adjust-filters/vibrance';
+export {default as vignette} from './shader-modules/adjust-filters/vignette';
 
 // glfx BLUR shader modules
-export {default as tiltShift} from './modules/blur-filters/tiltshift';
-export {default as triangleBlur} from './modules/blur-filters/triangleblur';
-export {default as zoomBlur} from './modules/blur-filters/zoomblur';
+export {default as tiltShift} from './shader-modules/blur-filters/tiltshift';
+export {default as triangleBlur} from './shader-modules/blur-filters/triangleblur';
+export {default as zoomBlur} from './shader-modules/blur-filters/zoomblur';
 
 // glfx FUN shader modules
-export {default as colorHalftone} from './modules/fun-filters/colorhalftone';
-export {default as dotScreen} from './modules/fun-filters/dotscreen';
-export {default as edgeWork} from './modules/fun-filters/edgework';
-export {default as hexagonalPixelate} from './modules/fun-filters/hexagonalpixelate';
-export {default as ink} from './modules/fun-filters/ink';
+export {default as colorHalftone} from './shader-modules/fun-filters/colorhalftone';
+export {default as dotScreen} from './shader-modules/fun-filters/dotscreen';
+export {default as edgeWork} from './shader-modules/fun-filters/edgework';
+export {default as hexagonalPixelate} from './shader-modules/fun-filters/hexagonalpixelate';
+export {default as ink} from './shader-modules/fun-filters/ink';
 
 // glfx WARP shader modules
-export {default as bulgePinch} from './modules/warp-filters/bulgepinch';
-export {default as swirl} from './modules/warp-filters/swirl';
+export {default as bulgePinch} from './shader-modules/warp-filters/bulgepinch';
+export {default as swirl} from './shader-modules/warp-filters/swirl';
 
 // experimental shader modules and passes
-export {default as _depth} from './experimental/modules/depth';
-export {default as _convolution} from './experimental/modules/convolution';
+export {default as _depth} from './experimental/shader-modules/depth';
+export {default as _convolution} from './experimental/shader-modules/convolution';
 export {default as _OutlinePass} from './experimental/passes/outline-pass';
 export {default as _SSAOPass} from './experimental/passes/ssao-pass';
